+++ start-repl.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+lein repl :start :host 0.0.0.0 :port 6666